Fix firefox memory issue
Synchronize geckoGeneration flags in VRShMem to avoid a bug that locks up
render thread.
geckoGeneration flags are probably not even needed after this
but I'll leave them in just to be on the safe side.
Rest of the diffs are just changes made to mozilla central branch over the years.